Seperate cli_main from ui_main
